requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==23.11.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
distro==1.9.0
filelock==3.18.0
Flask==3.0.2
flask-cors==6.0.1
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.4
openai==1.86.0
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
setuptools==78.1.0
sniffio==1.3.1
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
